Widen AIM/MERC hiring-availability subsystem UINT8->UINT16, add test mercs 255-260
Phase 6 flagged the AIM/MERC-site hiring path as still UINT8-capped even
after the profile ID system itself grew to 2048 slots (Phases 1-7). This
closes that gap and proves it out with six live test mercs above #255.

Widened UINT8 -> UINT16 (plain, not the ProfileID strong-typedef):
- Laptop/mercs.h/.cpp: CONDITION_FOR_MERC_AVAILABLE(_TEMP) fields,
  gubMercArray[], gubCurMercIndex, NUMBER_OF_MERCS, and every function
  signature carrying a profile ID across the M.E.R.C.-site hiring path.
- Laptop/aim.h/.cpp: AIM_AVAILABLE(_TEMP) fields, AimMercArray[],
  MAX_NUMBER_MERCS (same pattern, AIM site).
- Laptop/AimMembers.cpp: gbCurrentSoldier/gbCurrentIndex and two local
  hire-handler variables that would have aliased any AIM merc above #255
  to a low-numbered one.
- Laptop/AimSort.cpp: qsort() was called with sizeof(UINT8) against a
  UINT16 array - would have corrupted the array on first sort.
  QsortCompare's UINT8* reads fixed to match.
- Laptop/AimFacialIndex.cpp, Encyclopedia_Data_new.cpp, mercs Files.cpp:
  stale extern UINT8 re-declarations that no longer matched the real
  (now UINT16) definitions.
- Laptop/XML_ConditionsForMercAvailability.cpp, XML_AIMAvailability.cpp:
  parser casts were (UINT8) atol(...), truncating any profile ID above
  255 read from the availability XML files.
- Sentinel fixes (mercs.cpp x2, mercs Files.cpp x1, Soldier Profile.cpp):
  `!= 255` comparisons had to become `!= 0xFFFF`, since assigning -1 to
  a UINT16 field wraps to 65535, and AIMAvailability.xml's ~250 built-in
  empty slots are written as literal `<ProfilId>-1</ProfilId>`.

Added six AIM test mercs (profile indices 255-260, gamedir/Data-1.13/
TableData/MercProfiles.xml) with matching AIMAvailability.xml entries,
reusing profile #50's (Bubba's) face/voice/portrait art and background
text so nothing points at missing assets. Verified in a running game:
new game -> AIM -> all six (TestA-TestF) appear correctly, alongside the
existing roster, not aliased to a low-numbered profile.

Full JA2.exe build compiles and links clean.

Known, deliberately deferred (unchanged from Phase 6):
- Rebel Command's bit-packed mercProfileId is a separate bit-layout
  redesign, not a mechanical width bump.
- AIM Facial Index screen's mugshot-grid browser still paginates via a
  hardcoded 3-state cycle (0/40/80), capping that screen's display at
  120 mercs - needs its own pagination redesign later.

+++ Laptop/AimFacialIndex.cpp
@@ -17,7 +17,7 @@
 
 extern UINT8	gubCurrentSortMode; // symbol already defined in AimSort.cpp (jonathanl)
 extern UINT8	gubCurrentListMode; // symbol already declared globally in AimSort.cpp (jonathanl)
-extern UINT8	gbCurrentIndex;
+extern UINT16	gbCurrentIndex;	// Phase 6 follow-up: was UINT8 - must match AimMembers.cpp's real declaration
 
 
 UINT32		guiMugShotBorder;
@@ -367,7 +367,7 @@ void SelectMercFaceRegionCallBack(MOUSE_REGION * pRegion, INT32 iReason )
 	else if(iReason & MSYS_CALLBACK_REASON_LBUTTON_UP)
 	{
 		guiCurrentLaptopMode = LAPTOP_MODE_AIM_MEMBERS;
-		gbCurrentIndex = (UINT8) MSYS_GetRegionUserData( pRegion, 0 ) + START_MERC;
+		gbCurrentIndex = (UINT16) MSYS_GetRegionUserData( pRegion, 0 ) + START_MERC;
 	}
 	else if (iReason & MSYS_CALLBACK_REASON_RBUTTON_UP)
 	{
